# backend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Staff Scheduler API",
    "version": "1.0.0",
    "description": "Open-source workforce management REST API.\n\nAll non-public endpoints require a JWT obtained via `POST /api/v1/auth/login`, presented either as the httpOnly `token` cookie (set automatically at login) or as `Authorization: Bearer <token>`.\n\nBase URLs:\n- Canonical: `http://localhost:3001/api/v1`\n- Legacy (will be removed): `http://localhost:3001/api`\n\nEvery response follows the envelope format:\n- Success: `{ \"success\": true, \"data\": <T> }`\n- Error:   `{ \"success\": false, \"error\": { \"code\": \"ERROR_CODE\", \"message\": \"...\" } }`\n\nValidation errors return `VALIDATION_ERROR` with a `details` array of field-level messages.\nRate-limit errors return HTTP 429 with code `RATE_LIMIT_EXCEEDED`.",
    "contact": {
      "name": "Luca Ostinelli",
      "email": "ostinelliluca2@gmail.com",
      "url": "https://github.com/lucaosti/StaffScheduler"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Local development (canonical)"
    },
    {
      "url": "http://localhost:3001/api",
      "description": "Local development (legacy — will be removed)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT passed as Authorization: Bearer <token>. Alternative to the httpOnly cookie for non-browser clients."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "httpOnly JWT cookie set by POST /auth/login (primary mechanism for the web frontend)."
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code, e.g. VALIDATION_ERROR, NOT_FOUND, UNAUTHORIZED."
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "description": "Field-level validation errors (present on VALIDATION_ERROR).",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid JWT.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but lacking the required permission.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
    },
    "parameters": {
      "id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
...
    }
  },
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/directory/import-vcard": {
      "post": {
        "tags": [
          "Directory"
        ],
        "summary": "Bulk import from vCard",
        "description": "Admin only. Existing emails are skipped.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "vcf": {
                    "type": "string",
                    "minLength": 1
                  },
                  "defaultPassword": {
                    "type": "string",
                    "minLength": 8
                  }
                },
                "required": [
                  "vcf",
                  "defaultPassword"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inserted and skipped counts."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/directory/me": {
      "get": {
        "tags": [
          "Directory"
        ],
        "summary": "Own directory profile",
        "description": "Returns the caller's profile plus any custom fields.",
        "responses": {
          "200": {
            "description": "Profile."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/directory/users/{id}": {
      "get": {
        "tags": [
          "Directory"
        ],
        "summary": "User directory profile",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "Profile."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/directory/users/{id}/fields": {
      "put": {
        "tags": [
          "Directory"
        ],
        "summary": "Bulk upsert custom fields",
        "description": "Replaces all custom fields for a user. Manager only.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fields": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "key": {
                          "type": "string",
                          "minLength": 1
                        },
                        "value": {}
                      },
                      "required": [
                        "key",
                        "value"
                      ]
                    }
                  }
                },
                "required": [
                  "fields"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated profile."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/directory/users/{id}/fields/{key}": {
      "delete": {
        "tags": [
          "Directory"
        ],
        "summary": "Remove a custom directory field",
        "description": "Requires `user.manage`.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Field removed."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/directory/users/{id}/vcard": {
      "get": {
        "tags": [
          "Directory"
        ],
        "summary": "Single-user vCard 4.0",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "vCard.",
            "content": {
              "text/vcard": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/directory/vcard.vcf": {
      "get": {
        "tags": [
          "Directory"
        ],
        "summary": "Multi-user vCard archive",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "vCard archive.",
            "content": {
              "text/vcard": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user by ID",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "User object."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update user",
        "description": "Update email, name, role, or active status. Admins can update any user; regular users can update themselves (limited fields).",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "firstName": {
                    "type": "string",
                    "minLength": 1
                  },
                  "lastName": {
                    "type": "string",
                    "minLength": 1
                  },
                  "roleIds": {
                    "type": "array",
                    "items": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    }
                  },
                  "employeeId": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "position": {
                    "type": "string"
                  },
                  "hourlyRate": {
                    "type": "number",
                    "minimum": 0
                  },
                  "isActive": {
                    "type": "boolean"
                  },
                  "organizationName": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 120
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete user",
        "description": "Permanently deletes a user. Admin only.",
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}